A bunch of minor stuff

+++ src/parameters/ComputerParameter.hpp
@@ -35,7 +35,7 @@ namespace lapis {
 
 		Title _title{ "Computer-Specific Options" };
 
-		NumericTextBox _thread{ "Number of Threads:","thread", _defaultNThread(),
+		NumericTextBox _thread{ "Number of Threads:","thread", (double)_defaultNThread(),
 		"The number of threads to run Lapis on. Defaults to the number of cores on the computer" };
 		std::string _threadCmd = "thread";
 	};
